requirements.txt
altair
beautifulsoup4
folium
psycopg2-binary
pandas
requests
streamlit
psycopg2
python-dotenv
streamlit-folium
sqlalchemy
python-dateutil
